Separate post card components into:
RoadmapPostCard
BoardPostCard
FullPostCard
